+++ src/components/ui/badge.tsx
@@ -5,20 +5,20 @@ import { Slot } from "radix-ui"
 import { cn } from "@/lib/utils"
 
 const badgeVariants = cva(
-  "group/badge inline-flex h-5 w-fit shrink-0 items-center justify-center gap-1 overflow-hidden rounded-4xl border border-transparent px-2 py-0.5 text-xs font-medium whitespace-nowrap transition-all focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 has-data-[icon=inline-end]:pr-1.5 has-data-[icon=inline-start]:pl-1.5 aria-invalid:border-destructive aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 [&>svg]:pointer-events-none [&>svg]:size-3!",
+  "group/badge inline-flex h-6 w-fit shrink-0 items-center justify-center gap-1 overflow-hidden rounded-full border border-transparent px-2 py-0.5 text-xs font-medium whitespace-nowrap transition-all focus-visible:border-ring focus-visible:ring-3 focus-visible:ring-focus-ring has-data-[icon=inline-end]:pr-1.5 has-data-[icon=inline-start]:pl-1.5 aria-invalid:border-destructive aria-invalid:ring-focus-ring-destructive [&>svg]:pointer-events-none [&>svg]:size-3!",
   {
     variants: {
       variant: {
-        default: "bg-primary text-primary-foreground [a]:hover:bg-primary/80",
+        default: "bg-primary text-primary-foreground [a]:hover:bg-primary-hover",
         secondary:
-          "bg-secondary text-secondary-foreground [a]:hover:bg-secondary/80",
+          "bg-secondary text-secondary-foreground [a]:hover:bg-secondary-hover",
         destructive:
-          "bg-destructive-muted text-destructive-muted-foreground focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 [a]:hover:bg-destructive-muted/80",
+          "bg-destructive-muted text-destructive-muted-foreground [a]:hover:bg-destructive-muted-hover",
         success:
-          "bg-success-muted text-success-muted-foreground [a]:hover:bg-success-muted/80",
+          "bg-success-muted text-success-muted-foreground [a]:hover:bg-success-muted-hover",
         warning:
-          "bg-warning-muted text-warning-muted-foreground [a]:hover:bg-warning-muted/80",
-        info: "bg-info-muted text-info-muted-foreground [a]:hover:bg-info-muted/80",
+          "bg-warning-muted text-warning-muted-foreground [a]:hover:bg-warning-muted-hover",
+        info: "bg-info-muted text-info-muted-foreground [a]:hover:bg-info-muted-hover",
         outline:
           "border-border text-foreground [a]:hover:bg-muted [a]:hover:text-muted-foreground",
         ghost: